requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.4
executing==1.0.0
filelock==3.8.0
gunicorn==20.1.0
huggingface-hub==0.9.1
idna==3.3
ipykernel==6.15.2
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.5
jupyter-core==4.11.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.1
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.9.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stack-data==0.5.0
tokenizers==0.12.1
torch==1.12.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
